Remove PSR20 polyfill
I've decided against using the PSR20 polyfill as that might break when
the PSR interface does not match the here promoted version.

So currently it makes more sense to depend upon the stella-maris/clock
package than on the psr/clock package. As soon as the psr/clock is
published and matches this interface then the
StellaMaris\Clock\ClockInterface will extend Psr\CLock\CLockInterface
and therefore be immediately forward compatible.